initial attempt

+++ install/arch/do_create_systemd_process.sh
@@ -4,6 +4,11 @@ sudo cp /tmp/testd9t/testd9t.service /etc/systemd/system/testd9t.service
 
 binary_path=$(pwd)/main
 
+if [ ! -f $binary_path ]; then
+    echo "Binary not found"
+    exit 1
+fi
+
 sudo cp $binary_path /home/digiconvent/backend/main
 
 sudo systemctl enable testd9t
